@@ -20,12 +20,12 @@ dependencies:
   flutter:
     sdk: flutter
   equatable: ^2.0.3
-  flutter_bloc: ^7.0.1
-  flutter_modular: ^3.2.2+1
-  shared_preferences: ^2.0.6
-  url_launcher: ^6.0.6
-  msix: ^2.1.1
-  phosphor_flutter: ^1.2.0
+  flutter_bloc: ^7.2.0
+  flutter_modular: ^3.4.1+1
+  shared_preferences: ^2.0.7
+  url_launcher: ^6.0.9
+  msix: ^2.1.3
+  phosphor_flutter: ^1.3.1
 
 dev_dependencies:
   flutter_test:
